requirements.txt
mcp>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
fastapi>=0.68.0
uvicorn>=0.15.0
pyjwt>=2.3.0
redis>=4.0.0
pyyaml>=6.0.0
pillow>=9.0.0
prometheus-client>=0.12.0
psutil>=5.8.0
python-multipart>=0.0.5 